+++ lib/representable/definition.rb
@@ -50,7 +50,7 @@ module Representable
     end
 
     def default_for(value)
-      return self[:default] if skipable_nil_value?(value)
+      return self[:default] if skipable_empty_value?(value)
       value
     end
 
@@ -66,7 +66,6 @@ module Representable
       return true if array? and self[:render_empty] == false and value and value.size == 0  # TODO: change in 2.0, don't render emtpy.
       value.nil? and not self[:render_nil]
     end
-    alias_method :skipable_nil_value?, :skipable_empty_value? # TODO: remove in 1.9 .
 
     def create_binding(*args)
       self[:binding].call(self, *args)